compile-forms*: only eval if requested, line nums on eval errors.
Add two entry points to compile-forms*: compile-forms and
compile-and-eval-forms. The latter is used by load-file (and
:use-macros, :require-macros) to eval while loading the file (full
compilation and evaluation). The former is used by compile-file* to
compile but not evaluate the forms when doing ahead of time
compilation (i.e. the way that ClojureScript does compilation right
now via closure.clj).